Handle all json escape sequences using a byte-by-byte inspection. Also handle escaped hex codes, which are technically out of spec, but often sent in json data. This approach is significantly faster than previous methods.

This benchmark run shows that BenchmarkJSONReader runs approximately 2x as fast.

```
BEFORE

goos: darwin
goarch: amd64
pkg: github.com/qri-io/dataset/dsio
cpu: Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz
BenchmarkCBORWriterArrays-8    	   10000	    117795 ns/op
BenchmarkCBORWriterObjects-8   	    6321	    192148 ns/op
BenchmarkCBORReader-8          	    2022	    583629 ns/op
BenchmarkCSVWriterArrays-8     	    2220	    525497 ns/op
BenchmarkCSVWriterObjects-8    	    2248	    526928 ns/op
BenchmarkCSVReader-8           	    1894	    620847 ns/op
BenchmarkJSONWriterArrays-8    	    6708	    177715 ns/op
BenchmarkJSONWriterObjects-8   	    4374	    276743 ns/op
BenchmarkJSONReader-8          	     649	   1836256 ns/op
PASS
ok  	github.com/qri-io/dataset/dsio	11.293s

AFTER

goos: darwin
goarch: amd64
pkg: github.com/qri-io/dataset/dsio
cpu: Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz
BenchmarkCBORWriterArrays-8    	   10000	    119981 ns/op
BenchmarkCBORWriterObjects-8   	    6381	    193160 ns/op
BenchmarkCBORReader-8          	    2031	    581762 ns/op
BenchmarkCSVWriterArrays-8     	    2246	    528192 ns/op
BenchmarkCSVWriterObjects-8    	    2262	    525037 ns/op
BenchmarkCSVReader-8           	    1892	    619852 ns/op
BenchmarkJSONWriterArrays-8    	    6252	    187372 ns/op
BenchmarkJSONWriterObjects-8   	    4257	    286447 ns/op
BenchmarkJSONReader-8          	    1212	    982385 ns/op
PASS
ok  	github.com/qri-io/dataset/dsio	11.239s
```