@b5 2021-08-07
I'm a big fan of benchcmp. using your benchmarks from above:

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkCBORWriterArrays-8      117795        119981        +1.86%
BenchmarkCBORWriterObjects-8     192148        193160        +0.53%
BenchmarkCBORReader-8            583629        581762        -0.32%
BenchmarkCSVWriterArrays-8       525497        528192        +0.51%
BenchmarkCSVWriterObjects-8      526928        525037        -0.36%
BenchmarkCSVReader-8             620847        619852        -0.16%
BenchmarkJSONWriterArrays-8      177715        187372        +5.43%
BenchmarkJSONWriterObjects-8     276743        286447        +3.51%
BenchmarkJSONReader-8            1836256       982385        -46.50%
```

can confirm. is faster.